refactor(lib): split `Job.run()` into `start()` and `wait()`
This allows `Queue` to know when the `Job` has had its status set to
`Running`, so it can drop the lock and allow other threads to pick up a
`Job`. Before, it was possible for two threads to pick up the same `Job`
if there was no `stagger_duration` set!