refactor(lib): create field-free `Status` enum
There are a lot of places (though mostly tests) where I want to check
the `Status` of a `Job`, but some of the fields in `Running` and `Failed`
make it impossible to reasonably test for equality or derive `PartialEq`.
By creating a new, field-free version of `Status` (and a `From` impl for
converting from the existing `Status` enum), I can simplify a lot of
testing and application code.