fix(lib): race with `.kill()` after `Job.reset()`
I can't replicate this race in a test, so I'm really just operating off
of suspicion, a single instance of strange GUI behaviour, and my own
reasoning here.

I think that the way I'm handling `Job.reset()`s is a bit brittle when
it comes to resetting `Running` `Job`s. When I do that reset, I kill
the command running in the background. The thread that's then waiting
on that command will get a `Failed` status back. Because the killed
process always exits after the `Status` has been reset to `Queued`, it
"undoes" that resetting and `Running` jobs that have been reset will
always become `Failed` instead of `Queued`. To prevent that `Failed`
status from overriding the `Queued` one that we want, `Job.wait()`
has been re-checking the status to make sure that it's still `Running`
before updating it to either `Completed` or `Failed`; if it's not, then
we assume it's been reset and leave the status as-is.

I think the issue here is that `Job.reset()` instantly updates the
`Job`'s status to `Queued`, but there is a brief window of time before
the `Running` job is actually killed. Within this window, it's possible
for another worker to pick up and start the `Job`, setting its status
back to `Running`. Then the kill completes, sees that the `Job` is still
running, and updates the status to `Failed`. Now when the second worker
that picked up that `Job` finishes with it, it sees that the status is
no longer `Running` (instead it's `Failed`), so it *does not* update the
status to `Completed`. I think this is how I ended up with a `Running`
-> `Job.reset()` -> `Failed` chain of events when I was playing around
with the GUI.

To fix this, I'll introduce a new `Status::Stopping` that will prevent
`Failed` from being set, but also prevent other workers from picking
up the `Job` until the `.kill()` is complete. When the `.kill()` is
complete, the `Stopping` status can be reset to `Queued`.

This does, however, introduce a new problem: `on_update` is now being
called by `Job`s that are `clear()`ed from the `Queue`... I'll solve
this by introducing `Status::Abandoned` which won't trigger any
`on_change` call or return to a `Queued` state automatically. I'll pair
this with a new (crate-private) helper method: `Job.abandon()`.