fix(lib): more robust `assert_unpark_within_ms!()`
This version actually moves away from normal thread parking and allows
any number of `unpark()`s to preceed a `park()`. This is in contrast
to `thread::unpark()` which can only "queue" up to one unpark. This is
important because some operations will perform two `on_update` calls
back-to-back, and the second one is sometimes missed!