fix(lib): fix race with `Queue.cancel()`
If the `Queue` is cancelled whilst a `Job` is in the middle of spawning,
then it was possible for that `Job` to end up `Running` anyways and
escape killing. Now `Queue.cancel()` waits for any spawning `Job`s to
finish spawning (set their status to `Running`) before jumping in to
kill things.